+++ changelog.md
@@ -4,7 +4,8 @@
 * Removed CIM gRPC producers.
 * Added support for `EquivalentBranches` in `BusBranchNetworkCreator`.
 * `TownDetail` fields are now nullable.
-* Simplified `connect` and `connect_async` by refactoring optional settings (e.g. password authentication) to other functions
+* Simplified `connect` and `connect_async` by refactoring optional settings (e.g. password authentication) to other functions.
+  These functions are now deprecated.
 
 ##### New Features
 * Implemented database module for persisting to sqlite database.